backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
requests>=2.26.0
newspaper3k>=0.2.8
nltk>=3.6.5
numpy>=1.21.2
scikit-learn>=1.0.1
networkx>=2.6.3
sentence-transformers>=2.1.0
spacy>=3.1.3
tqdm>=4.64.0
urllib3<2.0
certifi>=2021.10.8
idna>=3.3
charset-normalizer>=2.0.0
beautifulsoup4>=4.10.0
lxml>=4.6.5
python-dateutil>=2.8.2
pytz>=2021.3
regex>=2021.8.3